trace-cmd listen: Always use /var/run even if installed in /usr/local
As /var/run is a tempfs directory, and does not persist between reboots, but
directories in /usr/local do, always use /var/run. It does not make sense to
use /usr/local/var/run, as there should only be one instance of a running
trace-cmd server anyway, even if there's more than one installation of it.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>