trace-cmd list: Add -A to show the agents that are registered
A guest can run trace-cmd agent to connect to the host to allow the host to
initiate tracing on the guest. Add a -A option to trace-cmd list that can be
run on the host to connect to the host server to list the guests that have
agents connect to the host.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>