trace-cmd: Extract handle_tsoffset_longopt()
This short patch extracts the function handle_tsoffset_longopt() from the big
switch in parse_record_options(). The purpose of doing that is to reduce
the size of that big switch, by moving into functions the code in cases containg
control flow statements. In addition, labeling that code with a function name,
makes it easier to understand and maintain.