trace-cmd: Move trace-hash-local.h in lib/trace-cmd/include
This patch moves trace-hash-local.h in lib/trace-cmd/include, as part of a
bigger restructuring plan of trace-cmd's code base. In this case, the new
directory in the "private headers" directory in the trace-cmd library. In the
next step, the source files of that library will be moved in lib/trace-cmd.