trace-cmd: Move plugin_* files in plugins/
This patch moves the source files of all plugins to a dedicated directory, as
part of a bigger restructuring plan of trace-cmd's code base.
The build system has been updated to support this new directory structure by
adding a separate Makefile in plugins/ and by making the parent Makefile
to invoke it.