trace-cmd: Make libtraceevent builable out-of-tree
This patch allows the libtraceevent to be buildable out-of-tree when the 'O'
variable is set at command like while invoking make. When this option is used,
the out-of-tree build will still fail (as it does currently) but with this
patch the final goal gets a little bit closer.

NOTE: the regular in-tree build of all targets clearly continues to work.