trace-cmd listen: Don't always change the group ownership
There's no reason to change the ownership of files and directories if they
are already the expect owner. This allows non-root processes to run
trace-cmd listen, when root sets up group ownership for users to do so. By
trying to always change the ownership of some paths, even when the ownership
is of the correct value, it can cause users that are in groups and have
the correct permissions to do the listening to fail.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>