trace-cmd: Add first part of trace-cmd record connecting to agents
Add initial code to have a trace-cmd record (manager) connect to trace-cmd
listen (server) and have the server pass the file descriptors of an active
agent to the manager. Then the manager can start communication with the
agent.

Note, the communication to the agent has not yet been implemented, only the
set up with the server has. trace-cmd record -A (the agent option) will die,
as the code to communicate to the agent does not exist yet.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>