[NOT A REAL PATCH]: my virt-server hacks + notes
This is just a flat-patch containing the hacks in virt-server's code
that helped me understand (partially) how it works, plus a .txt file
with some notes and questions.

My plan for the actual implementation of virt-server was about me writing
first a hacky proof-of-concept implementation [in this patch, the impl.
is clearly far from complete]. After that, the road for the real impl.
would be pretty much clear: it will be about re-thinking about the data
necessary data structures, using/creating the right functions (consistent
abstractions etc.), handling all the corner cases and preparing a series
of patches to tell the story behind that work, in the right way.